spell school rework part 1

+++ crawl-ref/source/mapdef.cc
@@ -5345,7 +5345,7 @@ bool item_list::parse_single_spec(item_spec& result, string s)
         const string owner = replace_all_of(strip_tag_prefix(s, "owner:"),
                                             "_", " ");
 
-        COMPILE_CHECK(SPTYP_LAST_SCHOOL < SHRT_MAX);
+        COMPILE_CHECK(SPTYP_LAST_SCHOOL < LONG_MAX);
         result.props[RANDBK_DISC1_KEY].get_short() = disc1;
         result.props[RANDBK_DISC2_KEY].get_short() = disc2;
         result.props[RANDBK_NSPELLS_KEY] = num_spells;
